Implement auth client-side
Adds Auth Service to handle login/logout, provide auth status, and manage the JWT in local storage.
Adds Auth Interceptor to attach the token to the request in Authorization headers when the user is logged in.
Adds Auth Guard function, applying it to City/Country edit/create routes.
Adds login component.
Navigates user on successful login to a return url if it was set as a querystring param when redirected by the auth guard.
Adds Login/Logout links to nav menu.
Hides City/Country creation link when user is not authenticated.